+++ plugins/govppmux/plugin_impl_govppmux.go
@@ -48,6 +48,13 @@ type Deps struct {
 	local.PluginInfraDeps // inject
 }
 
+// Config groups the configurable parameter of GoVpp
+type Config struct {
+	HealthCheckProbeInterval time.Duration `json:"health-check-probe-interval"`
+	HealthCheckReplyTimeout  time.Duration `json:"health-check-reply-timeout"`
+	HealthCheckThreshold     int           `json:"health-check-threshold"`
+}
+
 // FromExistingAdapter is used mainly for testing purposes.
 func FromExistingAdapter(vppAdapter adapter.VppAdapter) *GOVPPPlugin {
 	ret := &GOVPPPlugin{
@@ -69,6 +76,18 @@ func (plugin *GOVPPPlugin) Init() error {
 	}
 
 	plugin.PluginName = plugin.Deps.PluginName
+
+	cfg := defaultConfig()
+	found, err := plugin.PluginConfig.GetValue(&cfg)
+	if err != nil {
+		return err
+	}
+	if found {
+		govpp.SetHealthCheckProbeInterval(cfg.HealthCheckProbeInterval)
+		govpp.SetHealthCheckReplyTimeout(cfg.HealthCheckReplyTimeout)
+		govpp.SetHealthCheckThreshold(cfg.HealthCheckThreshold)
+		plugin.Log.Debug("Setting govpp parameters", cfg)
+	}
 
 	if plugin.vppAdapter == nil {
 		plugin.vppAdapter = vppapiclient.NewVppAdapter()
@@ -157,4 +176,13 @@ func (plugin *GOVPPPlugin) handleVPPConnectionEvents(ctx context.Context) {
 			return
 		}
 	}
+}
+
+func defaultConfig() Config {
+	c := Config{
+		HealthCheckProbeInterval: time.Second,
+		HealthCheckReplyTimeout:  100 * time.Millisecond,
+		HealthCheckThreshold:     1,
+	}
+	return c
 }